+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- update to git2 0.15 which uses gits safe.directory logic.
+- Update to git2 0.15 which uses gits `safe.directory` logic.
 
 ## [0.19.0] - 2022-05-11