+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - new editor: `html_sanitize`
 - `headers` site options to supply additional headers on requests
+- Support URLs with IP addresses
 
 ### Changed